target-repo/Concept-2-3D/concept3d/backend/requirement.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
pydantic

pymongo
wikipedia

transformers
torch
pillow
trimesh
scipy
diffusers
accelerate
safetensors
rembg
onnxruntime; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'
